refactor: move notes file
