+++ src/app/App/MainScreen/TopBar/index.tsx
@@ -2,8 +2,8 @@ import React, { FC } from 'react';
 import { FaXmark } from 'react-icons/fa6';
 import { cn } from '@bem-react/classname';
 
-import { INote, NoteId } from '../../../../core/Note';
 import { INotesController } from '../../../../core/storage/controllers/notes';
+import { INote, NoteId } from '../../../../types/notes';
 import { Icon } from '../../../components/Icon/Icon.bundle/common';
 import { TabsMenu } from '../../../components/TabsMenu';
